+++ src/binance/client/rest/responses/my_trade_entry.cr
@@ -21,10 +21,10 @@ module Binance::Responses
     getter symbol : String?
 
     @[JSON::Field(key: "id")]
-    getter trade_id : Int32
+    getter trade_id : Int64
 
     @[JSON::Field(key: "orderId")]
-    getter order_id : Int32
+    getter order_id : Int64
 
     @[JSON::Field(key: "price", converter: Binance::Converters::ToFloat)]
     getter price : Float64
